+++ R/count_na.R
@@ -106,7 +106,7 @@ count_na_helper <- function(x, cn) {
   }
 
   # now compute frequency, and return a proper data frame
-  frq_helper(values, sort.frq = "none", weight.by = NULL, cn = cn) %>%
+  frq_helper(values, sort.frq = "none", weight.by = NULL, cn = cn, auto.grp = NULL) %>%
     dplyr::select(-.data$val) %>%
     dplyr::filter(.data$label != "NA")
 }