+++ R/frq.R
@@ -4,12 +4,16 @@
 #' @description This function returns a frequency table of labelled vectors, as data frame.
 #'
 #' @param sort.frq Determines whether categories should be sorted
-#'          according to their frequencies or not. Default is \code{"none"}, so
-#'          categories are not sorted by frequency. Use \code{"asc"} or
-#'          \code{"desc"} for sorting categories ascending or descending order.
+#'   according to their frequencies or not. Default is \code{"none"}, so
+#'   categories are not sorted by frequency. Use \code{"asc"} or
+#'   \code{"desc"} for sorting categories ascending or descending order.
 #' @param weight.by Vector of weights that will be applied to weight all observations.
-#'          Must be a vector of same length as the input vector. Default is
-#'          \code{NULL}, so no weights are used.
+#'   Must be a vector of same length as the input vector. Default is
+#'   \code{NULL}, so no weights are used.
+#' @param auto.grp Numeric value, indicating the minimum amount of unique
+#'   values in a variable, at which automatic grouping into smaller  units
+#'   is done (see \code{\link{group_var}}). Default value for \code{auto.group}
+#'   is \code{NULL}, i.e. auto-grouping is off.
 #'
 #' @inheritParams descr
 #'
@@ -70,7 +74,7 @@
 #' @importFrom tibble add_row
 #' @importFrom sjlabelled get_label get_labels get_values copy_labels
 #' @export
-frq <- function(x, ..., sort.frq = c("none", "asc", "desc"), weight.by = NULL) {
+frq <- function(x, ..., sort.frq = c("none", "asc", "desc"), weight.by = NULL, auto.grp = NULL) {
 
   # get dot data
   x <- get_dot_data(x, dplyr::quos(...))
@@ -101,7 +105,8 @@ frq <- function(x, ..., sort.frq = c("none", "asc", "desc"), weight.by = NULL) {
             x = tmp[[1]],
             sort.frq = sort.frq,
             weight.by = weight.by,
-            cn = colnames(tmp)[1]
+            cn = colnames(tmp)[1],
+            auto.grp = auto.grp
           )
 
         attr(dummy, "group") <- get_grouped_title(x, grps, i, sep = "\n")
@@ -123,7 +128,8 @@ frq <- function(x, ..., sort.frq = c("none", "asc", "desc"), weight.by = NULL) {
           x = x[[i]],
           sort.frq = sort.frq,
           weight.by = weight.by,
-          cn = colnames(x)[i]
+          cn = colnames(x)[i],
+          auto.grp = auto.grp
         )
 
       # save data frame for return value
@@ -138,7 +144,11 @@ frq <- function(x, ..., sort.frq = c("none", "asc", "desc"), weight.by = NULL) {
 }
 
 
-frq_helper <- function(x, sort.frq, weight.by, cn) {
+#' @importFrom dplyr n_distinct full_join
+#' @importFrom stats na.omit xtabs na.pass
+#' @importFrom sjlabelled get_labels get_label
+#' @importFrom tibble add_row
+frq_helper <- function(x, sort.frq, weight.by, cn, auto.grp) {
   # remember type
   vartype <- var_type(x)
 
@@ -158,11 +168,19 @@ frq_helper <- function(x, sort.frq, weight.by, cn) {
     return(structure(class = "sjmisc.frq", list(mydat = mydat)))
   }
 
-  # get value labels (if any)
-  labels <- sjlabelled::get_labels(x, attr.only = T, include.values = "n", include.non.labelled = T)
-
   # get variable label (if any)
   varlab <- sjlabelled::get_label(x)
+
+  if (!is.null(auto.grp) && dplyr::n_distinct(x, na.rm = TRUE) >= auto.grp) {
+    gl <- group_labels(x, size = "auto", n = auto.grp)
+    x <- group_var(x, size = "auto", n = auto.grp)
+    gv <- sort(stats::na.omit(unique(x)))
+    names(gv) <- gl
+    attr(x, "labels") <- gv
+  }
+
+  # get value labels (if any)
+  labels <- sjlabelled::get_labels(x, attr.only = T, include.values = "n", include.non.labelled = T)
 
   # if we don't have variable label, use column name
   if (sjmisc::is_empty(varlab) && !sjmisc::is_empty(cn))
@@ -212,7 +230,7 @@ frq_helper <- function(x, sort.frq, weight.by, cn) {
 
     # replace NA with 0, for proper percentages, i.e.
     # missing values don't appear (zero counts)
-    mydat$frq <- suppressMessages(replace_na(mydat$frq, value = 0))
+    mydat$frq <- suppressMessages(sjmisc::replace_na(mydat$frq, value = 0))
   } else {
     # weight data?
     if (!is.null(weight.by)) {